chore: Moved 'connect with retries' helper from cmd/run to firefox/remote module
