+++ src/firefox/index.js
@@ -14,7 +14,7 @@ import {isErrorWithCode, UsageError, WebExtError} from '../errors';
 import {getPrefs as defaultPrefGetter} from './preferences';
 import {getManifestId} from '../util/manifest';
 import {createLogger} from '../util/logger';
-import {default as defaultFirefoxConnector, REMOTE_PORT} from './remote';
+import {connect as defaultFirefoxConnector, REMOTE_PORT} from './remote';
 // Import flow types
 import type {FirefoxConnectorFn} from './remote';
 import type {
